+++ deltafi-ui/src/components/FormattedBytes.vue
@@ -1,0 +1,44 @@
+<!--
+   DeltaFi - Data transformation and enrichment platform
+
+   Copyright 2022 DeltaFi Contributors <deltafi@deltafi.org>
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+-->
+
+<template>
+  <span v-tooltip.top="tooltip">{{ formattedBytes }}</span>
+</template>
+
+<script setup>
+import useUtilFunctions from "@/composables/useUtilFunctions";
+import { computed, defineProps } from "vue";
+
+const { formattedBytes: formatBytes } = useUtilFunctions();
+
+const props = defineProps({
+  bytes: {
+    type: Number,
+    required: true,
+  }
+});
+
+const tooltip = computed(() => {
+  return `${props.bytes.toLocaleString()} B`;
+});
+
+const formattedBytes = computed(() => {
+  return formatBytes(props.bytes);
+});
+
+</script>